podcast_manager/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.7
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-login==1.0
django-nose==1.4.7
django-ratelimit==4.1.0
django-redis==5.3.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
flower==2.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
humanize==4.8.0
hyperframe==5.2.0
idna==2.10
kavenegar==1.1.2
kombu==5.3.2
mock==5.1.0
model-bakery==1.14.0
nose==1.3.7
pika==1.3.2
Pillow==10.0.0
polls==0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gettext==5.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
rfc3986==1.5.0
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
South==1.0.2
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
